src/aers/utils/logging.py
from loguru import logger
import sys
def setup_logging(level: str = "INFO") -> None:
    logger.remove()
    logger.add(sys.stderr, level=level, colorize=true,
                format="green{time:YYYY-MM-DD HH:mm:ss.SSS} | "
                "<level>{level: <8}</level> | "
                "<cyan>{name}</cyan>:<cyan>{function}</cyan>:<cyan>{line}</cyan> | "
                "<level>{message}</level>")